Move DataContainer in Utils folder
