+++ src/Utils/CMakeLists.txt
@@ -1,6 +1,5 @@
 cmake_minimum_required(VERSION 3.13.0)
 
 target_sources(bonjourmonde PRIVATE
-    Inputs.cpp
-    InputEnums.cpp
+    DataContainer.cpp
 )
